minor workaround date selection bug
Unfortunately this only fixes the bug when having a meal selected
across date boundaries and swiping back from that. In other cases, as
in launching the app from the background without it being a new
launch, it doesn't appear that onAppear is called. I'm not sure if
there's anything else I can do here.

closes #61